support nested lists
this is a pretty big rewrite.  the main change was storing nested
messages as ListMut(*T) rather than ListMut(T).  this patch also
replaces ctx.reader with ctx.data (a buffer) and does pre-scanning of
all fields.  this may not be be necessary. i've added it following
protoc-c but and am leaving it in for now.

* plugin.pb.zig
  * simplified Message/EnumDescriptor.init() methods so they only need a
    type now
* protobuf-util.zig
  * deserialize() - pre-scan all fields. will likely remove this,
    preferring a reader interface rather then requiring an in memory
    buffer