split gen.zig into gen-{zig,c}.zig files
maintains the ability to generate c code but in its own file. now each
file implements a similar interface with a shared context and the
following methods:
  gen{Prelude,Enum,MessageTypedef,Message,EnumTest,MessageTest,Postlude}()

this isn't super clean but it does the job. now to switch to generating
c, just have to set `gen.output_format = .c`.

* includes some unrelated c code changes to match up c/zig structs sizes