protoc-gen-zig: act like a proper plugin
Previously we didn't send a CodeGeneratorResponse back to protoc and
unnecessarily copied the protoc request to stderr.  And because of
this, we were compelled to parse protoc cli args and write the files to
disk.  These rouge tactics are no longer necessary :^).

Now we send the file contents and name back to protoc as a
CodeGeneratorResponse which writes them to disk.  And we can skip
parsing the cli args.

This is a big relief which greatly simplifies the code.  Thank you
@SuperAuguste for helping me realize this change was necessary.